several simple refactorings
- move auth from default.py to auth.py
- rename errors to notfound
- drop basic templates (mytemplate.jinja2, layout.jinja2)
