+++ docs/tutorials/wiki2/src/authorization/tutorial/views/default.py
@@ -6,17 +6,9 @@ from pyramid.httpexceptions import (
     HTTPFound,
     HTTPNotFound,
     )
-from pyramid.view import (
-    view_config,
-    forbidden_view_config,
-    )
-from pyramid.security import (
-    remember,
-    forget,
-    )
+from pyramid.view import view_config
 
 from ..models import Page
-from ..security.default import USERS
 
 # regular expression used to find WikiWords
 wikiwords = re.compile(r"\b([A-Z]\w+[A-Z]+\w+)")
@@ -76,37 +68,3 @@ def edit_page(request):
         page=page,
         save_url=request.route_url('edit_page', pagename=pagename),
         )
-
-
-@view_config(route_name='login', renderer='templates/login.jinja2')
-@forbidden_view_config(renderer='templates/login.jinja2')
-def login(request):
-    login_url = request.route_url('login')
-    referrer = request.url
-    if referrer == login_url:
-        referrer = '/' # never use the login form itself as came_from
-    came_from = request.params.get('came_from', referrer)
-    message = ''
-    login = ''
-    password = ''
-    if 'form.submitted' in request.params:
-        login = request.params['login']
-        password = request.params['password']
-        if USERS.get(login) == password:
-            headers = remember(request, login)
-            return HTTPFound(location=came_from, headers=headers)
-        message = 'Failed login'
-
-    return dict(
-        message=message,
-        url=request.route_url('login'),
-        came_from=came_from,
-        login=login,
-        password=password,
-        )
-
-@view_config(route_name='logout')
-def logout(request):
-    headers = forget(request)
-    next_url = request.route_url('view_wiki')
-    return HTTPFound(location=next_url, headers=headers)
