+++ docs/tutorials/wiki2/src/basiclayout/tutorial/views/notfound.py
@@ -1,5 +1,7 @@
 from pyramid.view import notfound_view_config
 
+
 @notfound_view_config(renderer='../templates/404.jinja2')
 def notfound_view(request):
+    request.response.status = 404
     return {}
